Move DomainSpec to automation.py and add Mixin Class
