+++ homeassistant/components/button/trigger.py
@@ -2,9 +2,9 @@
 
 from homeassistant.const import STATE_UNAVAILABLE, STATE_UNKNOWN
 from homeassistant.core import HomeAssistant, State
+from homeassistant.helpers.automation import DomainSpec
 from homeassistant.helpers.trigger import (
     ENTITY_STATE_TRIGGER_SCHEMA,
-    DomainSpec,
     EntityTriggerBase,
     Trigger,
 )
