+++ homeassistant/components/cover/trigger.py
@@ -2,7 +2,8 @@
 
 from homeassistant.const import STATE_OFF, STATE_ON, STATE_UNAVAILABLE, STATE_UNKNOWN
 from homeassistant.core import HomeAssistant, State, split_entity_id
-from homeassistant.helpers.trigger import DomainSpec, EntityTriggerBase, Trigger
+from homeassistant.helpers.automation import DomainSpec
+from homeassistant.helpers.trigger import EntityTriggerBase, Trigger
 
 from .const import ATTR_IS_CLOSED, DOMAIN, CoverDeviceClass
