+++ homeassistant/components/schedule/trigger.py
@@ -2,8 +2,8 @@
 
 from homeassistant.const import STATE_OFF, STATE_ON, STATE_UNAVAILABLE, STATE_UNKNOWN
 from homeassistant.core import HomeAssistant, State
+from homeassistant.helpers.automation import DomainSpec
 from homeassistant.helpers.trigger import (
-    DomainSpec,
     EntityTransitionTriggerBase,
     Trigger,
     make_entity_target_state_trigger,
